2025 06 06 맹승주 Player 추가 수정

+++ Assets/Scripts/MapEditor/TileMapData.cs
@@ -12,8 +12,8 @@ public class TileData
 public class TileMapLayerData
 {
     public string layerName;
-    public int originX;
-    public int originY;
+    public float originX;
+    public float originY;
     public int width, height;
     public List<TileData> tiles = new List<TileData>();
 }
